Update README.md
Updated to include linking edits for Py_Initialize. Note: the AMPs method is still very experimental and should only really be used as a visualization for work on the amps research. There are no significant runtime or accuracy improvements. I'm not going to push an edited Makefile.common because it is outside of the directories we have here on github and only one line was added. -G