Calculator classes now in use
Part of completing #59. Each calculator now has its own class, which inherits from superclass Calculator. Calculator-specific code in LJ.cpp has been removed. This means that damping effects are no longer in use. Will open new issue of how and where damping should be implemented.

Calculators that depend on variables that may change between runs (ASE, pyamff) have stand-in values for the time being until ASE read-in is supported. Calculators that have intrinsic values that aren't in normal filetypes (LJ, Morse) currently have standin values as well, but there is not as clear of a way of moving forward to allow customization (I really don't want the solution to be putting a bunch of model-specific parameters in stdin).