Adding necessary library for AMFF
This file was excluded by gtignore, but is necessary. For now I am leaving the user with control over the PyAMFF folder (i.e. can/must run make in it), but for a more stable release it might be good to just put the compiled libraries in there and get rid of the makefiles.