Object oriented changes
After much work, all of the calculators have been moved out of LJ.cpp and into their own files. Going to keep this branch unmerged for a bit while I figure out what to do about all of the janky things that were being done to the LJ before and decide if they should be implemented.