Discovered and fixed ase.calculators.lj issue with import
When importing, there was some sort of path issue going on with my personal ubuntu 20.04 computer. The fix I found was to include duplicate imports, one in calculator.py and one in potentials.cpp. For the record, I don't like this solution, and I will be trying to figure out what is going on with the path at a future date. I would patch this just on my system, but as I would like the installation process to be as easy for newcomers as possible and because an extra import shouldn't cause issues, I am including it.

I also removed an unused import in calculator.py and added error handling for PyImport_Import, so if you have a python error it will send it to stdout.