Implemented David's boundary conditions
Put in some changes david made that capture atoms at the edge of the boundary if they are launched. I also included the boundary rules that were in LJ before creating the calculator class. We will want to scrutinize the original boundary rules to see what they are doing different and if they are necessary.