Added elements (#68)
Elements exist now in the haptic device. Each atom object has an atomicNumber attribute as well as a function that will find the mass and string element representation (e.g. "Sn" for tin). Force calculations take mass of the atom into account. Atoms display in a color that corresponds to the CPK color scheme. Resolves #68