Element support added with file read-in (Resolves #69)
You can now read in from con or poscar. Does not fix position yet, so weird scenario happens where first atom gets moved to (0,0,0) and the rest of the atoms stay at the spot in their respective box. There is a short-term fix in the commented out code around 677 of LJ that I will be putting in soon, but the best solution is to get boxes working (and get pbc while we're at it).