Created Demo Mode
Demo mode was added for research forum. If you run with "demo" instead of the normal calculator, you will get a modified lj that has special parameters for gold silver pair interactions. Do not run with other elements, or you will use the gold-silver potential for all of your calculations.